+++ src/business/index.ts
@@ -1,4 +1,3 @@
-export * as showmasterpiece from './showmasterpiece';
 export * as huarongdao from './huarongdao';
 export * as bubbleShooter from './bubbleShooter';
 export * as mikuFlick from './mikuFlick';