+++ src/business/showmasterpiece/logic/contexts/index.ts
@@ -1,2 +1,0 @@
-export { CartProvider } from './CartContext';
-export { useCartContext } from '../hooks';