+++ src/business/showmasterpiece/server/index.ts
@@ -1,9 +1,0 @@
-/**
- * ShowMasterpiece server exports
- */
-
-export * from './schema/masterpieces';
-export * from './schema/bookings';
-export * from './schema/popupConfig';
-export * from './schema/config';
-export * from './services';