+++ src/business/showmasterpiece/server/services/fileService.ts
@@ -1,1 +1,0 @@
-export { createOssFileConfigManagerFromEnv as getShowMasterpieceFileConfig } from '../../../../common/file/server';