+++ src/business/showmasterpiece/service/index.ts
@@ -1,4 +1,0 @@
-export * from './api';
-export * as webService from './web';
-export * as clientBusinessService from './client-business';
-export * as miniappService from './miniapp';