+++ src/business/showmasterpiece/service/web/index.ts
@@ -1,2 +1,0 @@
-export * from './client';
-export * as clientBusinessWebServices from '../client-business';