+++ src/business/showmasterpiece/ui/web/index.ts
@@ -1,3 +1,0 @@
-export { default as ShowMasterPiecesPage } from './pages/ShowMasterPiecesPage';
-export { default as ShowMasterPiecesConfigPage } from './pages/config/page';
-export { default as ShowMasterPiecesHistoryPage } from './pages/history/page';